move and rename src/websocket/instanceServices to src/api
