+++ webclient/src/api/RoomsService.tsx
@@ -1,4 +1,4 @@
-import { webClient } from "..";
+import { webClient } from "websocket";
 
 export default class RoomsService {
   static joinRoom(roomId) {
