+++ webclient/src/api/RouterService.tsx
@@ -1,4 +1,4 @@
-import { RouteEnum } from "../../types";
+import { RouteEnum } from "types";
 
 export class RouterService {
   resolveUrl(path, params) {
